Editor now allows for the game to be focused in and out by clicking on the game or pressing Q.
The editor takes the game and transposes it into a frame buffer object rendered to the editor game viewport.

Further features to be included revolve around editing the game structure directly through the editor and allowing hot reloading of the game code.

Next feature will be resizing of the editor game viewport.