More of a checkpoint, so the commit message here is just a copy of a summary conversation:
I've got it set up for testing and running right now, with the network actually initialized. My "testing" setup just modifies the pipeline so instead of Serialize data -> send packet (and then an independent Deserialize packets -> store data), it goes Serialize data in a stored buffer -> deserialize buffer and store/update data.

What i discovered is there is a serialization issue; it took a lot of debugging but I think i eventually figured it out
it looks like its as simple as that I neglected to account for the fact that the component data isn't stored in 8-bit integers
I mean, I kind of did account for it
the main way being that the serialization consistently uses "field sizes" for byte allocation
But there is one thing that doesnt use field sizes and the deserializer relies upon it being in the 3rd byte: the parent id
but the second byte isnt cast to a uint8_t, its a size_of. I will try casting it to a uint8_t but there is a decent chance that there will still be serialization issues; regardless I'll be done for now
If the other bits work because it dynamically allocates by field size and uses the same process to deserialize, then maybe itll work
oh yeah lmfao size_of returns a long long
so instead of the buffer being type, size, ID
it was type, size, 0, 0, 0, 0 ,0 ,0 ,0, ..., id
though that still doesnt quite explain the error I was getting, which was from get_components that the id didnt existy
but that doesnt make sense because id 0 defo does exist
but maybe caused a rolling ball of undefined behavior
Scoot — Today at 8:12 PM
okay that didnt quite fix it but did seem to fix my initial issue
ive also just had the realization that uint8_t on its own cannot hold a lot of the values we want it to
and i dont think our write_to_buffer or deserialization has a way of handling that right now
for instance it tried to store a value of 500:
Image
anyways
there will be more work to be done fs
at this point i wont even push unless you wanna work w it
I think the solution there might actually kind of be simple?
to just stretch it out
serialize into uint8_ts because i assume we have to with network shit
but when we serialize, evaluate the data into a uint32_t, and somehow split it raw
which i figure is actually easier done than said
then combine again when deserializing
and just do everything by multiples of 4
but maybe some brainstorming there could help find a better soln
Scoot — Today at 8:21 PM
Or possible we don't even need to use 8 bit integers at all and it's just set up that way to prevent bloat