Preparing for a refactor following new path forward on Sevrer-Client communication model. These changes reflect an effort to allow input-monitoring and server-authoritative object data.
BLUF: gse::input now stores a vector of all input mediums, allowing for a given input source to be accessed based on it's owner's ID.

For instance: The Server is trying to evaluate inputs from Player 4 (UUID = x). it can simply call gse::input::get_keyboard(x).

However, only the server actually needs to store mutliple inputs. That being said, by default, the "local" keyboard is actually keyboard 0, controller 0, mouse 0, etc; not associated with the player ID.

This is facilitated by a new "is_server" bool in gse::network, accessible by anything which can access its parent. When anything requiring player-dependent input is created, such as the player hook, it sets an "input id"; this is the local keyboard if on a client, but if it is the server, this will refer to the corresponding player keyboard. All of this runs without change to the update loop.

TLDR:

If I'm a client, nothing changed, really. I have a vector of inputs, which is necessary for server infrastructure without having a unique server input system, but I only keep track of one keyboard: Local/0.

If I am the server, when a new player connects, I simply add a new keyboard to g_keyboards with their player ID as the key. Once this is done, I simply update the keyboard with a corresponding packet. The actual game update loop otherwise does not change, automatically detecting which input keyboard it should use. (Not yet implemented)