Callback is implemented but not flushed out. Right now, the callback, map, and related functions ITSELF should be about ready for GSE as-is. Callback-related hashing functions have been added for use in maps.
The callbacks are added by ID and gse::input::key, providing per-entity callback registration. For instance, Player 1 (or a tank, horse, gun, etc) has the W key assigned to a provided lambda function. For repeated use, it could also be a pre-defined function. Test case was attempted for W/move forward; Ran into an out-of-range error. But so far, no reason to believe the callback itself didn't work, likely an oversight on the map which I will debug tomorrow.

 However, GSE's input handling system is fundamentally different from our vision of the callback-oriented input handling. So, the best I can implement right now is:

1) Add & define callback during a given hook's initialization loop.
2) During hook's update loop, check get_keyboard()[Given key].pressed is true
3) Retrieve the callback
4) .use() the callback.

My understanding of where we want to end:

1) Add & define callback during a given hook's initialization loop.
2) during a universal input monitoring loop, ACTIVATE corresponding callbacks.
3) during a given hook's update loop, use_activated_callbacks(owner_id) or something equivalent
4) re-set during universal input monitor loop.

The callbacks should easily support future-infrastructure changes like "activation", changes to the map assignment, etc. They are low-level and fairly abstract right now.