Completed per-object update loop multithreading. Investigated potential parallelization of collision detection but all options were either insignfiicant in execution time or could not execute in parallel.
Added an "accumulators" struct to the motion component to contain per-frame accumulators to allow for single-line reset. Added a position_correction accumulator to handle correction so that narrow phase does not directly modify the motion component when not needed. Similar implementation still needed for torque application.

Discovered that collision rotations are not processing properly, regardless of MPR issues. Set up a test case which effectively zeros out penetration depth to allow for easier debugging. Likely an issue in the final torque calculation, based off a similar symptom in the previous OpenGL branch implementation.